+++ lib/share_plus_dialog.dart
@@ -1,3 +1,5 @@
+/// A Flutter package that wraps share_plus to create sharing dialogs when
+/// share_plus is not available.
 library share_plus_dialog;
 
 export 'package:share_plus_dialog/src/share_dialog.dart';